requirements.txt
tweepy
pandas
nltk
gensim
scikit-learn
python-dotenv